+++ src/Configuration/CompletionConfiguration.cs
@@ -118,28 +118,6 @@ namespace TailwindCSSIntellisense.Configuration
             }
         }
 
-        private static bool IsHex(object value, out string hex)
-        {
-            var content = value.ToString().Trim('#').ToUpper();
-            var hexLetters = "ABCDEF";
-            if (content.All(c => char.IsNumber(c) || hexLetters.Contains(c)))
-            {
-                if (content.Length == 6 || content.Length == 8)
-                {
-                    hex = content.Substring(0, 6);
-                    return true;
-                }
-                else if (content.Length == 3)
-                {
-                    hex = content;
-                    return true;
-                }
-            }
-
-            hex = null;
-            return false;
-        }
-
         private bool GetDictionary(object value, out Dictionary<string, object> dict)
         {
             if (value is Dictionary<string, object> values)